refactor: extract study page state logic into custom hooks
- Create useStudyState hook to manage all study session logic
- Improve useStudyKeyboard with better loading state handling
- Simplify StudyPage component to focus only on rendering
- Add proper separation of concerns between UI and business logic
- Fix useCallback dependencies and improve performance
